+++ hutool-core/src/main/java/cn/hutool/core/util/StrUtil.java
@@ -350,10 +350,21 @@ public class StrUtil {
 	 *
 	 * @param args 被检查的对象,一个或者多个
 	 * @return 是否都不为空
-	 * @since 5.3.5
+	 * @since 5.3.6
 	 */
 	public static boolean isAllNotEmpty(CharSequence... args) {
 		return false == hasEmpty(args);
+	}
+
+	/**
+	 * 是否存都不为{@code null}或空对象或空白符的对象，通过{@link StrUtil#hasBlank(CharSequence...)} 判断元素
+	 *
+	 * @param args 被检查的对象,一个或者多个
+	 * @return 是否都不为空
+	 * @since 5.3.6
+	 */
+	public static boolean isAllNotBlank(CharSequence... args) {
+		return false == hasBlank(args);
 	}
 
 	/**